Retrofit added to project. Repository moved to a singleton instance. ProgressBar added to MovieListFragment and it is active while movie's list is loading.
