--- app/src/main/java/com/katdmy/android/myfirstkotlinapp/data/Movie.kt
@@ -1,19 +1,0 @@
-package com.katdmy.android.myfirstkotlinapp.data
-
-import android.os.Parcelable
-import kotlinx.parcelize.Parcelize
-
-@Parcelize
-data class Movie(
-    val id: Int,
-    val title: String,
-    val overview: String,
-    val poster: String,
-    val backdrop: String,
-    val ratings: Float,
-    val numberOfRatings: Int,
-    val minimumAge: Int,
-    val runtime: Int,
-    val genres: List<Genre>,
-    val actors: List<Actor>
-) : Parcelable
