+++ app/src/main/java/com/katdmy/android/myfirstkotlinapp/model/Actor.kt
@@ -1,0 +1,11 @@
+package com.katdmy.android.myfirstkotlinapp.model
+
+import android.os.Parcelable
+import kotlinx.parcelize.Parcelize
+
+@Parcelize
+data class Actor(
+    val id: Int,
+    val name: String,
+    val picture: String
+) : Parcelable
