+++ app/src/main/java/com/katdmy/android/myfirstkotlinapp/model/Genre.kt
@@ -1,0 +1,11 @@
+package com.katdmy.android.myfirstkotlinapp.model
+
+import android.os.Parcelable
+import kotlinx.parcelize.Parcelize
+
+@Parcelize
+data class Genre(val id: Int, val name: String) : Parcelable {
+    override fun toString(): String {
+        return name
+    }
+}
