+++ app/src/main/java/com/katdmy/android/myfirstkotlinapp/retrofit/Retrofit.kt
@@ -1,0 +1,48 @@
+package com.katdmy.android.myfirstkotlinapp.retrofit
+
+import com.jakewharton.retrofit2.converter.kotlinx.serialization.asConverterFactory
+import kotlinx.serialization.json.Json
+import okhttp3.MediaType.Companion.toMediaType
+import okhttp3.OkHttpClient
+import okhttp3.logging.HttpLoggingInterceptor
+import retrofit2.Retrofit
+import retrofit2.create
+
+object RetrofitClient {
+
+    const val MainServer = "https://api.themoviedb.org/3/"
+
+    val json = Json {
+        ignoreUnknownKeys = true
+    }
+
+    private val client = OkHttpClient().newBuilder()
+        .addInterceptor(HttpLoggingInterceptor().setLevel(HttpLoggingInterceptor.Level.BODY))
+        .build()
+
+    @Suppress("EXPERIMENTAL_API_USAGE")
+    private val retrofit: Retrofit = Retrofit.Builder()
+        .baseUrl(MainServer)
+        //.client(client)
+        .addConverterFactory(json.asConverterFactory("application/json".toMediaType()))
+        .build()
+
+    val tmdbApi: TmdbApi = retrofit.create()
+
+    /*
+    val retrofitClient: Retrofit.Builder by lazy {
+
+        val okhttpClient = OkHttpClient.Builder()
+
+        Retrofit.Builder()
+            .baseUrl(MainServer)
+            .client(okhttpClient.build())
+            .addConverterFactory(json.asConverterFactory("application/json".toMediaType()))
+    }
+
+    val tmdbApi: TmdbApi by lazy {
+        retrofitClient
+            .build()
+            .create(TmdbApi::class.java)
+    } */
+}
