+++ app/src/main/res/layout/fragment_movie_details.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@drawable/bg"
-    tools:context=".moviesdetails.MovieDetailsFragment">
+    tools:context=".view.MovieDetailsFragment">
 
     <androidx.core.widget.NestedScrollView
         android:layout_width="match_parent"
